fixed timestamp bug

+++ contracts/interfaces/ILoanManager.sol
@@ -6,18 +6,18 @@ import "../libraries/OfferLibrary.sol";
 import "../interfaces/ILoanManager.sol";
 
 interface ILoanManager {
-    function getLoan(uint160 loanId) external returns (LoanLibrary.Loan memory);
+    function getLoan(uint256 loanId) external returns (LoanLibrary.Loan memory);
 
     function repayLoan(
-        uint160 loanId,
+        uint256 loanId,
         uint256 interestPaid,
         uint256 principalPaid,
         uint256 collateralRetrieved
     ) external returns (bool);
 
-    // function liquidateLoan(uint160 loanId) external returns (bool);
+    // function liquidateLoan(uint256 loanId) external returns (bool);
 
-    function claimPrincipal(uint160 loanId) external;
+    function claimPrincipal(uint256 loanId) external;
 
-    function claimCollateral(uint160 loanId) external;
+    function claimCollateral(uint256 loanId) external;
 }
